tools/phantom-www/requirements.txt
# Phantom WWW Build System Requirements

# Core dependencies
markdown==3.5.1
python-frontmatter==1.0.1
Jinja2==3.1.2
PyYAML==6.0.1
markupsafe==2.1.3

# Optional for watch mode
watchdog==3.0.0

# Optional for optimization (production)
cssmin==0.2.0
rjsmin==1.2.1